Move files regarding outfit generation usecase-specific logic to the use_case.generate_outfit package. Modify Outfit and related classes to implement requiring an umbrella when it's raining
