+++ src/main/java/entity/Outfit.java
@@ -5,9 +5,11 @@ import java.util.Objects;
 
 public final class Outfit {
     private final Map<ClothingType, ClothingItem> clothingItems;
+    private final boolean isUmbrellaRequired;
 
-    public Outfit(Map<ClothingType, ClothingItem> clothingItems) {
+    public Outfit(Map<ClothingType, ClothingItem> clothingItems, boolean isUmbrellaRequired) {
         this.clothingItems = clothingItems;
+        this.isUmbrellaRequired = isUmbrellaRequired;
     }
 
     public Map<ClothingType, ClothingItem> getClothingItems() {
@@ -19,22 +21,23 @@ public final class Outfit {
     }
 
     @Override
-    public boolean equals(Object obj) {
-        if (obj == this) return true;
-        if (obj == null || obj.getClass() != this.getClass()) return false;
-        var that = (Outfit) obj;
-        return Objects.equals(this.clothingItems, that.clothingItems);
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        Outfit outfit = (Outfit) o;
+        return isUmbrellaRequired == outfit.isUmbrellaRequired && Objects.equals(clothingItems, outfit.clothingItems);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(clothingItems);
+        return Objects.hash(clothingItems, isUmbrellaRequired);
     }
 
     @Override
     public String toString() {
-        return "Outfit[" +
-                "clothingItems=" + clothingItems + ']';
+        return "Outfit{" +
+                "clothingItems=" + clothingItems +
+                ", isUmbrellaRequired=" + isUmbrellaRequired +
+                '}';
     }
-
 }
