+++ src/main/java/use_case/generate_outfit/RandomClothingItemSelectionStrategy.java
@@ -1,4 +1,6 @@
-package entity;
+package use_case.generate_outfit;
+
+import entity.ClothingItem;
 
 import java.util.List;
 import java.util.Random;
